+++ server/test-data/CPSC413/actual.json
@@ -1,58 +1,61 @@
-[
+{
+  "course": "CPSC 413",
+  "topic": " Design and Analysis of Algorithms I",
+  "assessments": [
     {
-        "course": "CPSC 413",
-        "topic": " Design and Analysis of Algorithms I",
-        "assessments": [
-            {
-                "name": "Mini-assignment 1",
-                "date": "2022-09-14T00:00:00.000",
-                "weight": "2.6"
-            },
-            {
-                "name": "Mini-assignment 2",
-                "date": "2022-09-26T00:00:00.000",
-                "weight": "2.6"
-            },
-            {
-                "name": "Mini-assignment 3",
-                "date": "2022-10-12T00:00:00.000",
-                "weight": "2.6"
-            },
-            {
-                "name": "Mini-assignment 4",
-                "date": "2022-11-02T00:00:00.000",
-                "weight": "2.6"
-            },
-            {
-                "name": "Mini-assignment 5",
-                "date": "2022-11-23T00:00:00.000",
-                "weight": "2.6"
-            },
-            {
-                "name": "Midterm",
-                "date": "2022-10-27T17:00:00.000",
-                "weight": "25"
-            },
-            {
-                "name": "Assignment 1",
-                "date": "2022-10-03T00:00:00.000",
-                "weight": "7.5"
-            },
-            {
-                "name": "Assignment 2",
-                "date": "2022-10-21T00:00:00.000",
-                "weight": "7.5"
-            },
-            {
-                "name": "Assignment 3",
-                "date": "2022-11-05T00:00:00.000",
-                "weight": "7.5"
-            },
-            {
-                "name": "Assignment 4",
-                "date": "2022-12-01T00:00:00.000",
-                "weight": "7.5"
-            }
-        ]
+      "name": "Mini-assignment 1",
+      "date": "2022-09-14T00:00:00.000",
+      "weight": "2.6"
+    },
+    {
+      "name": "Mini-assignment 2",
+      "date": "2022-09-26T00:00:00.000",
+      "weight": "2.6"
+    },
+    {
+      "name": "Mini-assignment 3",
+      "date": "2022-10-12T00:00:00.000",
+      "weight": "2.6"
+    },
+    {
+      "name": "Mini-assignment 4",
+      "date": "2022-11-02T00:00:00.000",
+      "weight": "2.6"
+    },
+    {
+      "name": "Mini-assignment 5",
+      "date": "2022-11-23T00:00:00.000",
+      "weight": "2.6"
+    },
+    {
+      "name": "Midterm",
+      "date": "2022-10-27T17:00:00.000",
+      "weight": "25"
+    },
+    {
+      "name": "Assignment 1",
+      "date": "2022-10-03T00:00:00.000",
+      "weight": "7.5"
+    },
+    {
+      "name": "Assignment 2",
+      "date": "2022-10-21T00:00:00.000",
+      "weight": "7.5"
+    },
+    {
+      "name": "Assignment 3",
+      "date": "2022-11-05T00:00:00.000",
+      "weight": "7.5"
+    },
+    {
+      "name": "Assignment 4",
+      "date": "2022-12-01T00:00:00.000",
+      "weight": "7.5"
+    },
+    {
+      "name": "Final Exam",
+      "date": "TB",
+      "weight": "32"
     }
-]
+  ]
+}
